renamed and refactored route for clarity
