+++ event_routes/get_semester.py
